+++ packages/browser/rollup.config.js
@@ -1,18 +1,18 @@
-import babel from 'rollup-plugin-babel';
+import typescript from 'rollup-plugin-typescript2';
 import uglify from 'rollup-plugin-uglify';
 import replace from 'rollup-plugin-replace';
 import resolve from 'rollup-plugin-node-resolve';
 
 const config = {
-  entry: 'src/index.js',
+  entry: 'src/index.ts',
   moduleName: 'HickoryBrowser',
   sourceMap: true,
   globals: {
     '@hickory/root': 'HickoryRoot'
   },
   plugins: [
-    babel({
-      exclude: 'node_modules/**'
+    typescript({
+      useTsconfigDeclarationDir: true
     }),
     replace({
       "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV)
