+++ packages/browser/tests/unit/browser.spec.ts
@@ -1,3 +1,4 @@
+import "jest";
 import Browser from '../../src';
 import { jsdom } from 'jsdom';
